Ritveer/ritveer_project/requirements.txt
# Core LangChain & Orchestration
langchain
langgraph
langchain-community
langchain-ollama

# Web Server for Hooks (e.g., Twilio)
fastapi
uvicorn

# Utilities
python-dotenv  # For loading .env files
pyyaml         # For loading policy.yml
pydantic-settings # ADD THIS LINE for robust settings management

# Database & Cache
redis          # For Redis connection
psycopg2-binary # For PostgreSQL connection

# External Services
twilio
razorpay
shiprocket